requirements.txt
Django>=5.2.5
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
requests>=2.31.0
